Add ability to generate a clusterid field for k8s
When running multiple k8s clusters, it's possible for multiple agents to
have the same IP address. To differentiate between these agents, we can send
a clusterid which is the UID of the default namespace. The namespace is
guaranteed to exist, so we can use it as a unique identifier of a cluster.

To accommodate delegation, the non-delegated agents can query and send ONLY
the clusterid. All other k8s info is empty in the protobuf for those agents.