Move k8s delegation logic to libsinsp
For k8s delegation, we need to query the k8s api server for nodes. This was
previously done in the analyzer by a separate object, but that work is folded
into the existing nodes handler. If this agent determines that it is delegated,
it will create the remaining handlers and populate the protobuf.

This first piece adds the delegation plumbing and dynamically creating
handlers when the node handler detects delegation.